requirements.txt
google-generativeai>=0.8.0
gspread>=6.0.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
requests>=2.31.0
